Add form hooked up

+++ server/db/models/review.js
@@ -7,7 +7,7 @@ const Review = db.define('review', {
         allowNull: false
     },
     rating:{
-        type: Sequelize.FLOAT(1,5),
+        type: Sequelize.INTEGER,
         allowNull: false
     },
 })
